chore(backend): moved /:userId/files endpoint to the UsersController
