--- apps/zero-api/src/users/users-files.controller.ts
@@ -1,45 +1,0 @@
-import {
-  ClassSerializerInterceptor,
-  Controller,
-  ForbiddenException,
-  Get,
-  Param,
-  ParseIntPipe,
-  UseFilters,
-  UseInterceptors,
-  UsePipes,
-  ValidationPipe
-} from '@nestjs/common';
-import { NoDataInterceptor } from '../interceptors/NoDataInterceptor';
-import { PrismaClientExceptionFilter } from '../exception-filters/PrismaClientExceptionFilter';
-import { UsersService } from './users.service';
-import { FilesService } from '../files/files.service';
-import { ApiBearerAuth, ApiOkResponse, ApiTags } from '@nestjs/swagger';
-import { User } from './decorators/user.decorator';
-import { FileMetadataDto } from '../files/dto/file-metadata.dto';
-import { UserDto } from './dto/user.dto';
-import { UserRole } from '@prisma/client';
-
-@Controller('users/:userId/files')
-@UsePipes(ValidationPipe)
-@UseInterceptors(ClassSerializerInterceptor, NoDataInterceptor)
-@UseFilters(PrismaClientExceptionFilter)
-export class UsersFilesController {
-  constructor(
-    private readonly usersService: UsersService,
-    private readonly filesService: FilesService
-  ) {}
-
-  @Get()
-  @ApiBearerAuth('access-token')
-  @ApiTags('users')
-  @ApiOkResponse({ type: [FileMetadataDto] })
-  async getUserFilesMetadata(
-    @User() user: UserDto,
-    @Param('userId', new ParseIntPipe()) userId: number
-  ): Promise<FileMetadataDto[]> {
-    if (user.id !== userId && !user.roles.includes(UserRole.admin)) throw new ForbiddenException();
-
-    return this.filesService.getUserFilesMetadata(userId);
-  }
-}
